refactor: #22 NotFountException 리팩토링
