+++ src/main/java/com/techeeresc/tab/domain/post/service/PostService.java
@@ -8,7 +8,7 @@ import com.techeeresc.tab.domain.post.entity.Post;
 import com.techeeresc.tab.domain.post.entity.QPost;
 import com.techeeresc.tab.domain.post.repository.PostQueryDslRepository;
 import com.techeeresc.tab.domain.post.repository.PostRepository;
-import com.techeeresc.tab.global.exception.customexception.RequestNotFoundException;
+import com.techeeresc.tab.global.exception.exceptionclass.RequestNotFoundException;
 import com.techeeresc.tab.global.status.StatusCodes;
 import com.techeeresc.tab.global.status.StatusMessage;
 import lombok.RequiredArgsConstructor;
@@ -22,6 +22,7 @@ public class PostService implements PostQueryDslRepository {
     private final PostRepository POST_REPOSITORY;
     private final PostMapper POST_MAPPER;
     private final JPAQueryFactory JPA_QUERY_FACTORY;
+    private final int NULL_SIZE = 0;
 
     @Transactional
     public Post insertPost(PostCreateRequestDto postCreateRequestDto) {
@@ -30,13 +31,19 @@ public class PostService implements PostQueryDslRepository {
 
     @Transactional
     public List<Post> findAllPost() {
-        return POST_REPOSITORY.findAll();
+        List<Post> posts = POST_REPOSITORY.findAll();
+        try {
+            isPostExistedByList(posts);
+            return posts;
+        } catch(NullPointerException exception) {
+            throw new RequestNotFoundException(StatusMessage.NOT_FOUND.getStatusMessage(), StatusCodes.NOT_FOUND);
+        }
     }
 
     @Transactional
     public Post updatePost(PostUpdateRequestDto postUpdateRequestDto) {
         try {
-            Post post = isPostExisted(postUpdateRequestDto.getId());
+            Post post = isPostExistedById(postUpdateRequestDto.getId());
             return post.updatePost(postUpdateRequestDto);
         } catch(NullPointerException exception) {
             throw new RequestNotFoundException(StatusMessage.NOT_FOUND.getStatusMessage(), StatusCodes.NOT_FOUND);
@@ -46,7 +53,7 @@ public class PostService implements PostQueryDslRepository {
     @Transactional
     public Post increaseLikeNumbers(Long id) {
         try {
-            Post post = isPostExisted(id);
+            Post post = isPostExistedById(id);
             return post.increaseLikeNumbers(post.getLikeNumbers());
         } catch(NullPointerException exception) {
             throw new RequestNotFoundException(StatusMessage.NOT_FOUND.getStatusMessage(), StatusCodes.NOT_FOUND);
@@ -56,7 +63,7 @@ public class PostService implements PostQueryDslRepository {
     @Transactional
     public List<Post> deletePost(Long id) {
         try {
-            Post post = isPostExisted(id);
+            Post post = isPostExistedById(id);
             POST_REPOSITORY.deleteById(post.getId());
         } catch (NullPointerException exception) {
             throw new RequestNotFoundException(StatusMessage.NOT_FOUND.getStatusMessage(), StatusCodes.NOT_FOUND);
@@ -68,7 +75,7 @@ public class PostService implements PostQueryDslRepository {
     @Transactional
     public Post findPostByIdAndIncreaseViews(Long id) {
         try {
-            Post post = isPostExisted(id);
+            Post post = isPostExistedById(id);
             post = increaseViews(post);
             return post;
         } catch(NullPointerException exception) {
@@ -79,7 +86,7 @@ public class PostService implements PostQueryDslRepository {
     @Transactional
     public Post findPostById(Long id) {
         try {
-            Post post = isPostExisted(id);
+            Post post = isPostExistedById(id);
             return post;
         } catch(NullPointerException exception) {
             throw new RequestNotFoundException(StatusMessage.NOT_FOUND.getStatusMessage(), StatusCodes.NOT_FOUND);
@@ -93,6 +100,9 @@ public class PostService implements PostQueryDslRepository {
         try {
             List<Post> postSearchResults = JPA_QUERY_FACTORY.selectFrom(qPost)
                     .where(qPost.title.contains(word)).fetch();
+
+            isPostExistedByList(postSearchResults);
+
             return postSearchResults;
         } catch (NullPointerException exception) {
             throw new RequestNotFoundException(StatusMessage.NOT_FOUND.getStatusMessage(), StatusCodes.NOT_FOUND);
@@ -103,10 +113,16 @@ public class PostService implements PostQueryDslRepository {
         return post.increaseViews(post.getViews());
     }
 
-    private Post isPostExisted(Long id) {
+    private Post isPostExistedById(Long id) {
         Post post = POST_REPOSITORY.findById(id).orElseThrow(() ->
                 new NullPointerException());
 
         return post;
+    }
+
+    private void isPostExistedByList(List<Post> postSearchResults) {
+        if (postSearchResults.size() == NULL_SIZE) {
+            throw new NullPointerException();
+        }
     }
 }
