+++ src/main/java/com/techeeresc/tab/global/exception/response/ErrorResponse.java
@@ -2,7 +2,9 @@ package com.techeeresc.tab.global.exception.response;
 
 import com.techeeresc.tab.global.status.StatusCodes;
 import com.techeeresc.tab.global.status.StatusMessage;
+import lombok.Getter;
 
+@Getter
 public class ErrorResponse {
     private int errorCode;
     private String message;
